Extract generic FragmentStateAdapter
